feat: implement routes and views for admin panels
