+++ src/admin/views/stockTable.ejs
@@ -34,7 +34,7 @@
 							<button type="button" class="px-2 py-1 text-xs bg-gray-200 rounded hover:bg-gray-3" title="Aumentar quantidade">+</button>
 						</div>
 					</td>
-					<td class="px-6 py-3 flex gap-2">
+					<td class="px-6 py-3 flex">
 						<button type="button" class="hover:text-gray flex justify-center items-center" title="Bloquear">
 							<span class="material-symbols-outlined text-base">block</span>
 						</button>
