+++ src/coupons/views/couponsTable.ejs
@@ -1,6 +1,6 @@
 <div class="flex flex-col w-full gap-10 mt-9 mb-15 px-32">
 	<div class="flex justify-between align-middle">
-		<h1 class="font-semibold text-[20px]">Meus Cupons (1)</h1>
+		<h1 class="font-semibold text-[20px]">Meus Cupons (2)</h1>
 		<div class="flex items-center gap-2">
 			<h1 class="text-gray">Filtrar: </h1>
 			<select
