+++ src/generic/routes/defaultRouter.ts
@@ -6,7 +6,7 @@ import orderRouter from "../../order/router/orderRouter";
 import couponsRouter from "../../coupons/router/couponsRouter";
 import { BookControllerFactory } from "../../books/factories/BookControllerFactory";
 import bookRouter from "../../books/routes/bookRouter";
-import stockRouter from "../../stock/router/orderRouter";
+import adminRouter from "../../admin/routes/adminRouter";
 
 const defaultRouter = Router();
 const bookController = new BookControllerFactory().make();
@@ -32,40 +32,12 @@ defaultRouter.get("/shopping-cart", async (req: Request, res: Response) => {
 	});
 });
 
-defaultRouter.get("/admin/dashboard", async (req: Request, res: Response) => {
-	const books = await bookController.getAll(req, res);
-
-	const labels = ["Jan", "Fev", "Mar", "Abr", "Mai", "Jun"];
-
-	const salesHistory = (books ?? []).map((book: any) => {
-		const sales = labels.map(() => Math.floor(Math.random() * 50) + 10); // vendas aleatórias
-		return {
-			label: book.name,
-			data: sales,
-			fill: false,
-			borderColor: `#${Math.floor(Math.random()*16777215).toString(16)}`, // cor aleatória
-			tension: 0.3,
-		};
-	});
-
-	res.render("dashboard", {
-		title: "Dashboard - Grafico de linha de vendas",
-		currentHeaderTab: "profile",
-		layout: "detailsLayout",
-		currentUrl: "dashboard",
-		isAdmin: true,
-		books,
-		labels,
-		salesHistory
-	});
-});
-
 defaultRouter.use("/clients", clientRouter);
 defaultRouter.use("/addresses", addressRouter);
 defaultRouter.use("/cards", cardRouter);
 defaultRouter.use("/orders", orderRouter);
 defaultRouter.use("/coupons", couponsRouter);
 defaultRouter.use("/books", bookRouter);
-defaultRouter.use("/admin/stock", stockRouter);
+defaultRouter.use("/admin", adminRouter);
 
 export default defaultRouter;
