+++ src/order/router/orderRouter.ts
@@ -20,26 +20,6 @@ orderRouter.get("/", (req: Request, res: Response) => {
 	});
 });
 
-orderRouter.get("/admin", (req: Request, res: Response) => {
-	res.render("ordersTable", {
-		title: "Meus Pedidos",
-		currentHeaderTab: "profile",
-		layout: "detailsLayout",
-		currentUrl: "orders",
-		isAdmin: true
-	});
-});
-
-orderRouter.get("/admin/exchange-orders", (req: Request, res: Response) => {
-	res.render("exchangeOrdersTable", {
-		title: "Pedidos de Troca",
-		currentHeaderTab: "profile",
-		layout: "detailsLayout",
-		currentUrl: "exchange-orders",
-		isAdmin: true,
-	});
-});
-
 orderRouter.get("/:id", async (req: Request, res: Response) => {
 	const books = await bookController.getAll(req, res);
 	res.render("orderDetails", {
@@ -50,19 +30,6 @@ orderRouter.get("/:id", async (req: Request, res: Response) => {
 		books: books,
 		currentUrl: "orders",
 		isAdmin: false
-	});
-});
-
-orderRouter.get("/admin/:id", async (req: Request, res: Response) => {
-	const books = await bookController.getAll(req, res);
-	res.render("orderDetails", {
-		title: "Detalhes do Pedido",
-		currentHeaderTab: "profile",
-		layout: "detailsLayout",
-		isNewOrder: false,
-		books: books,
-		currentUrl: "orders",
-		isAdmin: true
 	});
 });
