+++ src/order/views/ordersTable.ejs
@@ -33,7 +33,7 @@
 					</span>
 				</td>
 				<td class="px-6 py-3 flex gap-2">
-					<a href="<%= isAdmin ? "/orders/admin/1" : "/orders/1" %>" class="hover:text-gray flex justify-center items-center">
+					<a href="<%= isAdmin ? "/admin/orders/1" : "/orders/1" %>" class="hover:text-gray flex justify-center items-center">
 						<span class="material-symbols-outlined text-base">visibility</span>
 					</a>
 				</td>
@@ -47,7 +47,7 @@
 					</span>
 				</td>
 				<td class="px-6 py-3 flex gap-2">
-					<a href="<%= isAdmin ? "/orders/admin/1" : "/orders/1" %>" class="hover:text-gray flex justify-center items-center">
+					<a href="<%= isAdmin ? "/admin/orders/1" : "/orders/1" %>" class="hover:text-gray flex justify-center items-center">
 						<span class="material-symbols-outlined text-base">visibility</span>
 					</a>
 				</td>
