--- src/stock/router/orderRouter.ts
@@ -1,27 +1,0 @@
-import { Router, Request, Response } from "express";
-import ConfigDynamicPaths from "../../generic/helpers/ConfigDynamicPaths";
-import path from "path";
-import { BookControllerFactory } from "../../books/factories/BookControllerFactory";
-
-const stockRouter = Router();
-stockRouter.use(
-	ConfigDynamicPaths.configViewsPath(path.join(__dirname, "../views"))
-);
-
-const bookController = new BookControllerFactory().make();
-
-stockRouter.get("/", async (req: Request, res: Response) => {
-	const books = await bookController.getAll(req, res);
-
-	res.render("stockTable", {
-		title: "Estoque",
-		currentHeaderTab: "profile",
-		layout: "detailsLayout",
-		currentUrl: "stock",
-		isAdmin: true,
-		books: books
-	});
-});
-
-
-export default stockRouter;
